refactor: 타입 관련 중복 파일 제거 및 초기 Product 폴더 제거
