+++ src/refactoring/models/cart/actions/index.ts
@@ -1,4 +1,2 @@
-export * from "./add";
-export * from "./remove";
 export * from "./update";
 export * from "./calculate";
