+++ src/refactoring/models/cart/types.ts
@@ -1,4 +1,15 @@
-import { Product } from "../product";
+export interface Discount {
+  quantity: number;
+  rate: number;
+}
+
+export interface Product {
+  id: string;
+  name: string;
+  price: number;
+  stock: number;
+  discounts: Discount[];
+}
 
 export interface CartItem {
   product: Product;
