+++ src/refactoring/models/index.ts
@@ -1,2 +1,1 @@
 export * from "./cart";
-export * from "./product";
