--- src/refactoring/models/product/helpers.ts
@@ -1,7 +1,0 @@
-export const validateProductQuantity = (
-  currentQuantity: number,
-  newQuantity: number
-) => {
-  const validatedQuantity = Math.max(0, Math.min(newQuantity, currentQuantity));
-  return validatedQuantity;
-};
