+++ src/refactoring/utils/index.ts
@@ -1,2 +1,6 @@
-export * from "./cart";
-// export * from "./product";
+import { CartItem, Product } from "../models";
+
+export const getRemainingStock = (product: Product, cart: CartItem[]) => {
+  const cartItem = cart.find((item) => item.product.id === product.id);
+  return product.stock - (cartItem?.quantity || 0);
+};
