--- src/types.ts
@@ -1,24 +1,0 @@
-export interface Product {
-  id: string;
-  name: string;
-  price: number;
-  stock: number;
-  discounts: Discount[];
-}
-
-export interface Discount {
-  quantity: number;
-  rate: number;
-}
-
-export interface CartItem {
-  product: Product;
-  quantity: number;
-}
-
-export interface Coupon {
-  name: string;
-  code: string;
-  discountType: 'amount' | 'percentage';
-  discountValue: number;
-}
